pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tablesense2"
version = "0.1.0"
description = "TableSense-style spreadsheet table detection system"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Your Name" }]
dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "openpyxl>=3.1.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "matplotlib>=3.8.0",
  "wandb>=0.15.0",
  "pyyaml>=6.0",
  "tqdm>=4.65.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]

[tool.pytest.ini_options]
pythonpath = ["."]

